+++ apps/native-component-list/src/screens/Camera/CameraScreen.tsx
@@ -446,7 +446,7 @@ export default class CameraScreen extends React.Component<object, State> {
           mode={this.state.mode}
           mute={this.state.mute}
           zoom={this.state.zoom}
-          videoQuality="2160p"
+          videoQuality="1080p"
           onMountError={this.handleMountError}
           barcodeScannerSettings={{
             barcodeTypes: ['qr', 'pdf417'],