+++ packages/expo-camera/ios/CameraViewModule.swift
@@ -159,7 +159,9 @@ public final class CameraViewModule: Module, ScannerResultHandler {
       }
 
       OnViewDidUpdateProps { view in
-        view.initCamera()
+        Task {
+          await view.initCamera()
+        }
       }
 
       AsyncFunction("resumePreview") { view in
@@ -180,17 +182,21 @@ public final class CameraViewModule: Module, ScannerResultHandler {
         #if targetEnvironment(simulator) // simulator
         try takePictureForSimulator(self.appContext, view, options, promise)
         #else // not simulator
-        view.takePicture(options: options, promise: promise)
+        Task {
+          await view.takePicture(options: options, promise: promise)
+        }
         #endif
-      }.runOnQueue(.main)
+      }
 
       AsyncFunction("record") { (view, options: CameraRecordingOptions, promise: Promise) in
         #if targetEnvironment(simulator)
         throw Exceptions.SimulatorNotSupported()
         #else
-        view.record(options: options, promise: promise)
+        Task {
+          await view.record(options: options, promise: promise)
+        }
         #endif
-      }.runOnQueue(.main)
+      }
 
       AsyncFunction("stopRecording") { view in
         #if targetEnvironment(simulator)
@@ -198,7 +204,7 @@ public final class CameraViewModule: Module, ScannerResultHandler {
         #else
         view.stopRecording()
         #endif
-      }.runOnQueue(.main)
+      }
     }
 
     AsyncFunction("launchScanner") { (options: VisionScannerOptions?) in