+++ packages/expo-file-system/ios/Next/FileSystemDirectory.swift
@@ -7,6 +7,7 @@ internal final class FileSystemDirectory: FileSystemPath {
   }
 
   func validateType() throws {
+    try validatePermission(.read)
     var isDirectory: ObjCBool = false
     if FileManager.default.fileExists(atPath: url.path, isDirectory: &isDirectory) {
       if !isDirectory.boolValue {
@@ -17,6 +18,7 @@ internal final class FileSystemDirectory: FileSystemPath {
 
   func create() throws {
     try validateType()
+    try validatePermission(.write)
     guard !exists else {
       throw UnableToCreateDirectoryException("directory already exists")
     }
@@ -28,6 +30,12 @@ internal final class FileSystemDirectory: FileSystemPath {
   }
 
   var exists: Bool {
+    do {
+      try validatePermission(.read)
+    } catch {
+      return false
+    }
+
     var isDirectory: ObjCBool = false
     if FileManager.default.fileExists(atPath: url.path, isDirectory: &isDirectory) {
       return isDirectory.boolValue
@@ -37,6 +45,7 @@ internal final class FileSystemDirectory: FileSystemPath {
 
   // Internal only function
   func listAsRecords() throws -> [[String: Any]] {
+    try validatePermission(.read)
     var contents: [[String: Any]] = []
 
     let items = try FileManager.default.contentsOfDirectory(at: url, includingPropertiesForKeys: nil)