+++ packages/expo-updates/src/Updates.ts
@@ -8,7 +8,6 @@ import {
   UpdateFetchResult,
   UpdatesCheckAutomaticallyValue,
   UpdatesLogEntry,
-  UpdatesNativeStateMachineContext,
 } from './Updates.types';
 
 /**
@@ -81,6 +80,11 @@ export const isEmergencyLaunch = ExpoUpdates.isEmergencyLaunch;
  * what failed during initialization.
  */
 export const emergencyLaunchReason = ExpoUpdates.emergencyLaunchReason;
+
+/**
+ * Number of milliseconds it took to launch.
+ */
+export const launchDuration = ExpoUpdates.launchDuration;
 
 /**
  * This will be true if the currently running update is the one embedded in the build,
@@ -296,52 +300,4 @@ export async function fetchUpdateAsync(): Promise<UpdateFetchResult> {
     };
   }
   return result;
-}
-
-/**
- * @hidden
- */
-export function clearUpdateCacheExperimentalAsync(_sdkVersion?: string) {
-  console.warn(
-    "This method is no longer necessary. `expo-updates` now automatically deletes your app's old bundle files!"
-  );
-}
-
-/**
- * @hidden
- */
-export function transformNativeStateMachineContext(
-  originalNativeContext: UpdatesNativeStateMachineContext & {
-    latestManifestString?: string;
-    downloadedManifestString?: string;
-    lastCheckForUpdateTimeString?: string;
-    rollbackString?: string;
-  }
-): UpdatesNativeStateMachineContext {
-  const nativeContext = { ...originalNativeContext };
-  if (nativeContext.latestManifestString) {
-    nativeContext.latestManifest = JSON.parse(nativeContext.latestManifestString);
-    delete nativeContext.latestManifestString;
-  }
-  if (nativeContext.downloadedManifestString) {
-    nativeContext.downloadedManifest = JSON.parse(nativeContext.downloadedManifestString);
-    delete nativeContext.downloadedManifestString;
-  }
-  if (nativeContext.lastCheckForUpdateTimeString) {
-    nativeContext.lastCheckForUpdateTime = new Date(nativeContext.lastCheckForUpdateTimeString);
-    delete nativeContext.lastCheckForUpdateTimeString;
-  }
-  if (nativeContext.rollbackString) {
-    nativeContext.rollback = JSON.parse(nativeContext.rollbackString);
-    delete nativeContext.rollbackString;
-  }
-  return nativeContext;
-}
-
-/**
- * @hidden
- */
-export async function getNativeStateMachineContextAsync(): Promise<UpdatesNativeStateMachineContext> {
-  const nativeContext = await ExpoUpdates.getNativeStateMachineContextAsync();
-  return transformNativeStateMachineContext(nativeContext);
 }